chore: reorganize examples directory structure
- Consolidate examples into a single examples/ directory
- Move basic example from example/ to examples/basic/
- Rename conflict_resolution to conflict-resolution for consistency
- Group utility files in examples/utils/
- Create comprehensive README files for each section
- Update module paths and imports
- Add deprecation notes for examples that need API updates

This creates a cleaner, more discoverable structure:
- examples/basic/ - foundational notes app with dashboard (needs API updates)
- examples/conflict-resolution/ - advanced counter sync with conflict resolution
- examples/utils/ - testing utilities and helper functions
