+++ examples/basic/dashboard.png
