+++ examples/basic/dashboard/dashboard.go
