+++ examples/basic/demo.go
