+++ examples/basic/go.mod
@@ -1,4 +1,4 @@
-module github.com/c0deZ3R0/go-sync-kit/example
+module github.com/c0deZ3R0/go-sync-kit/examples/basic
 
 go 1.24.4
 
@@ -6,4 +6,4 @@ require github.com/c0deZ3R0/go-sync-kit v0.5.0
 
 require github.com/mattn/go-sqlite3 v1.14.30 // indirect
 
-replace github.com/c0deZ3R0/go-sync-kit => ../
+replace github.com/c0deZ3R0/go-sync-kit => ../..
