+++ examples/basic/metrics/metrics.go
