+++ examples/conflict-resolution/README.md
