+++ examples/conflict-resolution/client/conflict_resolver.go
