+++ examples/conflict-resolution/client/event.go
