+++ examples/conflict-resolution/client/resolver.go
