+++ examples/conflict-resolution/counter.go
