+++ examples/conflict-resolution/go.mod
