+++ examples/conflict-resolution/main.go
