+++ examples/conflict-resolution/resolver.go
