+++ examples/conflict-resolution/start_demo.sh
