+++ examples/go.mod
@@ -1,7 +1,13 @@
-module go-sync-kit-examples
+module github.com/c0deZ3R0/go-sync-kit/examples
+
+// This module is deprecated - individual examples now have their own modules.
+// See:
+//   - examples/basic/
+//   - examples/conflict-resolution/
+//   - examples/utils/
 
 go 1.24.4
 
 replace github.com/c0deZ3R0/go-sync-kit => ../
 
-require github.com/c0deZ3R0/go-sync-kit v0.6.0
+require github.com/c0deZ3R0/go-sync-kit v0.7.1
