+++ examples/utils/go.mod
@@ -1,0 +1,7 @@
+module github.com/c0deZ3R0/go-sync-kit/examples/utils
+
+go 1.24.4
+
+require github.com/c0deZ3R0/go-sync-kit v0.7.1
+
+replace github.com/c0deZ3R0/go-sync-kit => ../..
