+++ examples/utils/memory_store.go
